feat: Add printId function for printing string or number IDs
This commit adds a new function called `printId` which takes a parameter `id` of type `stringOrNumber` (a union type of string and number) and logs it to the console. This function is useful for printing IDs in a flexible manner. The function is demonstrated by calling it with both a string and a number in the code changes.