feat: Add address insertion functionality
This commit adds the functionality to insert address information for users in the `insertData` function of the `insert_index.ts` file. It modifies the existing code to insert the city, country, street, and pincode values into the `addresses` table along with the user's ID. This addition enhances the functionality of the code by allowing the storage of address information for users.