feat: Add getUserWithAddress function
This commit adds the getUserWithAddress function to the join_index.ts file. The function retrieves a user's information along with their address from the database by performing a join operation between the users and addresses tables. This addition enhances the functionality of the code by allowing the retrieval of user and address data in a single query.